# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code4Me V2 API",
    "description": "The complete API for Code4Me V2",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/user/delete": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete User",
        "description": "Delete the authenticated user's account together with all data collected about it.\n\nArgs:\n    delete_data (bool): Ignored; kept so existing clients keep working.\n    auth_token (str): Auth token provided in cookies to authenticate the user.\n    app (App): Dependency-injected app instance with DB and Redis access.\n\nReturns:\n    JsonResponseWithStatus: A success or error response depending on the outcome.",
        "operationId": "delete_user_api_user_delete_delete",
        "parameters": [
          {
            "name": "delete_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Deprecated and ignored: deleting an account always erases its data.",
              "default": false,
              "title": "Delete Data"
            },
            "description": "Deprecated and ignored: deleting an account always erases its data."
          },
          {
            "name": "auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserDeleteResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidOrExpiredAuthToken"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserBlockedError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/chat/delete/{chat_id}": {
      "delete": {
        "tags": [
          "Chat"
        ],
        "summary": "Delete Chat",
        "description": "Delete a specific chat by its ID.\nValidates that the user has access to the chat through their session and project tokens.",
        "operationId": "delete_chat_api_chat_delete__chat_id__delete",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          },
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          },
          {
            "name": "project_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteChatSuccessResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredProjectToken"
                    }
                  ],
                  "title": "Response 401 Delete Chat Api Chat Delete  Chat Id  Delete"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoAccessToGetQueryError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryNotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteChatError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteChatError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to delete chat"
          }
        },
        "type": "object",
        "title": "DeleteChatError",
        "description": "Error response for chat deletion failure."
      },
      "DeleteChatSuccessResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Chat deleted successfully"
          }
        },
        "type": "object",
        "title": "DeleteChatSuccessResponse",
        "description": "Acknowledgement response for successful chat deletion."
      },
      "DeleteUserBlockedError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "This account owns research studies or agent profiles, so it cannot delete itself."
          }
        },
        "type": "object",
        "title": "DeleteUserBlockedError"
      },
      "DeleteUserDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "User is deleted successfully."
          }
        },
        "type": "object",
        "title": "DeleteUserDeleteResponse"
      },
      "DeleteUserError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to delete the user!"
          }
        },
        "type": "object",
        "title": "DeleteUserError"
      },
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "InvalidOrExpiredAuthToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Authentication token not found! You are not authenticated or your token has expired. Login before you can perform this action."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredAuthToken"
      },
      "InvalidOrExpiredProjectToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Invalid or expired project token. Acquire a session token first."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredProjectToken"
      },
      "InvalidOrExpiredSessionToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session not found! You are not authenticated or your session has expired. Login before you can perform this action."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredSessionToken"
      },
...
      "NoAccessToGetQueryError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "You do not have permission to access this query."
          }
        },
        "type": "object",
        "title": "NoAccessToGetQueryError"
      },
...
      "QueryNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Query not found"
          }
        },
        "type": "object",
        "title": "QueryNotFoundError"
      },
...
      "UserNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "User not found"
          }
        },
        "type": "object",
        "title": "UserNotFoundError"
      },
...
    }
  }
}